Move dashboard setting handling into own command classes
The dashboards don't really belong to the user command.
