+++ ng/CMakeLists.txt
@@ -66,6 +66,7 @@ set (NG_JS_SRC_FILES
      ${NG_SRC_DIR}/src/gmp/commands/credentials.js
      ${NG_SRC_DIR}/src/gmp/commands/cves.js
      ${NG_SRC_DIR}/src/gmp/commands/cvsscalculator.js
+     ${NG_SRC_DIR}/src/gmp/commands/dashboards.js
      ${NG_SRC_DIR}/src/gmp/commands/dfncert.js
      ${NG_SRC_DIR}/src/gmp/commands/entities.js
      ${NG_SRC_DIR}/src/gmp/commands/entity.js
