+++ ng/src/gmp/gmp.js
@@ -34,6 +34,7 @@ import './commands/credentials.js';
 import './commands/cpes.js';
 import './commands/cves.js';
 import './commands/cvsscalculator.js';
+import './commands/dashboards.js';
 import './commands/dfncert.js';
 import './commands/feedstatus.js';
 import './commands/filters.js';
